requirements.txt
ms-api>=0.11.100
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pyyaml>=6.0
aiohttp>=3.9.0
